backend/requirements.txt
# Core Django dependencies
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# PDF processing
PyMuPDF>=1.23.0

# NLP and ML dependencies
nltk>=3.8.0
scikit-learn>=1.3.0
numpy>=1.24.0

# AI and LLM dependencies
google-generativeai>=0.3.0

# IPFS and blockchain
requests>=2.31.0
web3>=6.0.0

# File handling
Pillow>=10.0.0

